refactor(editor): isolate gui screen contract
